src/pyproject.toml
[tool.poetry]
name = 'project'
version = '0.1.0'
description = ''
authors = ['paulinapalma <paulina.anna.palma@gmail.com>']
readme = 'README.md'

[tool.poetry.dependencies]
python = '^3.12'
fastapi = '^0.114.2'
uvicorn = '^0.30.6'
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
python-multipart = "^0.0.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}



[tool.poetry.group.dev.dependencies]
mypy = '^1.11.2'

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'
